# backend/postman/specs/openapi.yaml
openapi: 3.0.3
info:
  title: MSM — Sistema MIP de Productividad y Reconocimiento Comercial
  description: |
    API del sistema MSM (Sistema MIP) para medir, gestionar y reconocer el
    desempeño de la fuerza de ventas farmacéutica.

    **Autenticación**: Bearer JWT. Usar `POST /auth/login` con form-data
    (`username` + `password`) para obtener el `access_token`.

    **Roles**: ADMIN · PRESIDENCIA · DIR_COMERCIAL · GERENTE_PRODUCTIVIDAD ·
    GERENTE_DISTRITO · GERENTE_MARCA · REPRESENTANTE_MEDICO · CONSULTA
  version: 1.0.0
  contact:
    email: segura.moises@gmail.com
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /dims/preview:
    post:
      tags: [Importación DIMs]
      summary: Preview de hojas en el Excel de catálogos
      description: |
        Sube el archivo DIM_MIP_FINAL.xlsx y retorna las hojas detectadas
        con su mapeo a tabla del sistema. No escribe en BD.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Archivo Excel (.xlsx) con hojas DIM_*
      responses:
        '200':
          description: Hojas detectadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  hojas:
                    type: array
                    items:
                      type: object
                      properties:
                        nombre_hoja:    { type: string }
                        tabla_sistema:  { type: string }
                        label:          { type: string }
                        filas:          { type: integer }
                        columnas:       { type: array, items: { type: string } }
                        reconocida:     { type: boolean }
                        orden:          { type: integer }
                  total_hojas:       { type: integer }
                  hojas_reconocidas: { type: integer }
        '400': { description: Formato de archivo inválido }
        '401': { description: No autenticado }
        '403': { description: Sin permiso }
  /dims/importar:
    post:
      tags: [Importación DIMs]
      summary: Importar hojas seleccionadas a la BD
      description: |
        Hojas disponibles: DIM_PAIS, DIM_LINEA, DIM_GERENTE, DIM_RM,
        DIM_INDICADOR, DIM_INDICADOR_TABLA, DIM_CICLO, DIM_MES,
        DIM_RECONOCIMIENTO, DIM_META_INDICADOR, DIM_CATEGORIA_DESEMPENO,
        DIM_KPI_DASHBOARD. Los registros ya existentes se omiten (no duplica).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, hojas]
              properties:
                file:
                  type: string
                  format: binary
                hojas:
                  type: string
                  description: 'JSON array de nombres de hojas. Ej: ["DIM_PAIS","DIM_LINEA"]'
                  example: '["DIM_PAIS","DIM_LINEA","DIM_GERENTE","DIM_RM","DIM_INDICADOR","DIM_INDICADOR_TABLA","DIM_CICLO","DIM_MES"]'
      responses:
        '200':
          description: Resultados por hoja
          content:
            application/json:
              schema:
                type: object
                properties:
                  resultados:
                    type: array
                    items:
                      type: object
                      properties:
                        nombre_hoja: { type: string }
                        label:       { type: string }
                        exitoso:     { type: boolean }
                        insertados:  { type: integer }
                        omitidos:    { type: integer }
                        errores:     { type: integer }
                        mensaje:     { type: string }
                  total_insertados: { type: integer }
                  total_omitidos:   { type: integer }
                  total_errores:    { type: integer }
        '400': { description: Archivo inválido o hojas mal formateadas }

  # ─── ETL ──────────────────────────────────────────────────────────────────
...